refactor(quickstart): split the scene into entry, shared, and per-variant folders
Quickstart.tsx becomes a variant dispatcher only. Components move to
shared/ (page chrome, product cards, modals, wizard progress), full-page/
(the test arm: guides, publications, installation prompt), and
simplified/ (the test2 arm: focused install, run status, hero answer).
One component per file; bodies are unchanged.
