+++ frontend/src/scenes/quickstart/full-page/PublicationSkeletonCard.tsx
@@ -1,0 +1,16 @@
+import { LemonSkeleton } from '@posthog/lemon-ui'
+
+import { LemonCard } from 'lib/lemon-ui/LemonCard'
+
+export function PublicationSkeletonCard(): JSX.Element {
+    return (
+        <LemonCard
+            hoverEffect={false}
+            className="flex flex-col gap-2 p-3 h-full rounded-lg border-transparent shadow-sm"
+        >
+            <LemonSkeleton className="w-full h-24 rounded" />
+            <LemonSkeleton className="w-3/4 h-4" />
+            <LemonSkeleton className="w-full h-3" />
+        </LemonCard>
+    )
+}
