+++ frontend/src/scenes/quickstart/full-page/QuickstartInstallationProgress.tsx
@@ -1,81 +1,11 @@
-import { useActions, useMountedLogic, useValues } from 'kea'
+import { useActions } from 'kea'
 
 import { IconCheckCircle, IconPullRequest, IconWarning } from '@posthog/icons'
 
-import { useFeatureFlag } from 'lib/hooks/useFeatureFlag'
 import { cn } from 'lib/utils/css-classes'
-import { activeCloudRunLogic, CloudRunHandle } from 'scenes/onboarding/shared/wizard-sync/activeCloudRunLogic'
-import { finishedLocalRunLogic } from 'scenes/onboarding/shared/wizard-sync/finishedLocalRunLogic'
 import { currentTaskLabel, pipClass, stepCounts, syncHeadline } from 'scenes/onboarding/shared/wizard-sync/helpers'
-import {
-    InstallationProgress,
-    installationProgressLogic,
-    progressFromFinishedLocalRun,
-} from 'scenes/onboarding/shared/wizard-sync/installationProgressLogic'
-import { wizardActiveSessionDetectorLogic } from 'scenes/onboarding/shared/wizard-sync/wizardActiveSessionDetectorLogic'
+import { InstallationProgress } from 'scenes/onboarding/shared/wizard-sync/installationProgressLogic'
 import { wizardSyncUiLogic } from 'scenes/onboarding/shared/wizard-sync/wizardSyncUiLogic'
-
-import { ProductKey } from '~/queries/schema/schema-general'
-
-interface QuickstartWizardProgressProps {
-    children: (progress: InstallationProgress) => React.ReactNode
-    fallback: React.ReactNode
-}
-
-function QuickstartCloudProgress({
-    activeCloudRun,
-    children,
-}: {
-    activeCloudRun: CloudRunHandle
-    children: QuickstartWizardProgressProps['children']
-}): JSX.Element {
-    const { installationProgress } = useValues(
-        installationProgressLogic({
-            mode: 'cloud',
-            runId: activeCloudRun.runId,
-            taskId: activeCloudRun.taskId,
-        })
-    )
-
-    return <>{children(installationProgress)}</>
-}
-
-function QuickstartLocalProgress({ children }: { children: QuickstartWizardProgressProps['children'] }): JSX.Element {
-    const { installationProgress } = useValues(installationProgressLogic({ mode: 'local' }))
-
-    return <>{children(installationProgress)}</>
-}
-
-function QuickstartLocalProgressDetector({ children, fallback }: QuickstartWizardProgressProps): JSX.Element {
-    useMountedLogic(wizardActiveSessionDetectorLogic)
-    const { hasActiveSession } = useValues(wizardActiveSessionDetectorLogic)
-    const { finishedLocalRun } = useValues(finishedLocalRunLogic)
-
-    if (hasActiveSession) {
-        return <QuickstartLocalProgress>{children}</QuickstartLocalProgress>
-    }
-    if (finishedLocalRun) {
-        return <>{children(progressFromFinishedLocalRun(finishedLocalRun))}</>
-    }
-    return <>{fallback}</>
-}
-
-export function QuickstartWizardProgress({ children, fallback }: QuickstartWizardProgressProps): JSX.Element {
-    const syncEnabled = useFeatureFlag('ONBOARDING_WIZARD_SYNC', 'test')
-    const { activeCloudRun } = useValues(activeCloudRunLogic)
-    const { finishedLocalRun } = useValues(finishedLocalRunLogic)
-
-    if (activeCloudRun) {
-        return <QuickstartCloudProgress activeCloudRun={activeCloudRun}>{children}</QuickstartCloudProgress>
-    }
-    if (syncEnabled) {
-        return <QuickstartLocalProgressDetector fallback={fallback}>{children}</QuickstartLocalProgressDetector>
-    }
-    if (finishedLocalRun) {
-        return <>{children(progressFromFinishedLocalRun(finishedLocalRun))}</>
-    }
-    return <>{fallback}</>
-}
 
 // The chip's leading glyph: the page signals live activity with pulsing dots (not spinners),
 // so working phases reuse that idiom and terminal phases get a small static icon.
@@ -138,15 +68,5 @@ export function QuickstartInstallationProgress({ progress }: { progress: Install
                 </span>
             </button>
         </div>
-    )
-}
-
-export function isQuickstartProductInstalling(
-    productKey: ProductKey,
-    installationProgress?: InstallationProgress
-): boolean {
-    return (
-        productKey === ProductKey.PRODUCT_ANALYTICS &&
-        (installationProgress?.phase === 'connecting' || installationProgress?.phase === 'running')
     )
 }
