+++ frontend/src/scenes/quickstart/full-page/SubsectionHeader.tsx
@@ -1,0 +1,3 @@
+export function SubsectionHeader({ title }: { title: string }): JSX.Element {
+    return <h3 className="text-sm font-semibold mb-3">{title}</h3>
+}
