+++ frontend/src/scenes/quickstart/shared/SectionHeader.tsx
@@ -1,0 +1,8 @@
+export function SectionHeader({ title, subtitle }: { title: string; subtitle?: string }): JSX.Element {
+    return (
+        <div className="mb-4">
+            <h2 className="text-lg font-semibold mb-0">{title}</h2>
+            {subtitle && <p className="text-secondary mb-0 mt-1">{subtitle}</p>}
+        </div>
+    )
+}
