+++ frontend/src/scenes/quickstart/shared/modals/QuickstartModals.tsx
@@ -1,0 +1,11 @@
+import { TaskGuidanceModal } from './TaskGuidanceModal'
+import { ToolSetupModal } from './ToolSetupModal'
+
+export function QuickstartModals({ installationComplete }: { installationComplete: boolean }): JSX.Element {
+    return (
+        <>
+            <TaskGuidanceModal />
+            <ToolSetupModal installationComplete={installationComplete} />
+        </>
+    )
+}
